Fixed it!
I'm dumb. It was counting the outer edges as... edges. That's fixed. Did
some messing around with the image; you know how to draw circles on the
original image (viscircles) but I wasn't sure how to get it to work.

Just need to get it to scan for clusters and draw circles, now.